Organize components into per-component directories
