+++ assets/js/components/index.js
@@ -1,9 +1,9 @@
-import './settings.js';
-import './theme-slider.js';
-import './card.js';
-import './copy-button.js';
-import './resource-card.js';
-import './icon-card.js';
-import './color-changer.js';
-import './theme-control.js';
-import './navigation.js';
+import './settings/settings.js';
+import './theme-slider/theme-slider.js';
+import './card/card.js';
+import './copy-button/copy-button.js';
+import './resource-card/resource-card.js';
+import './icon-card/icon-card.js';
+import './color-changer/color-changer.js';
+import './theme-control/theme-control.js';
+import './navigation/navigation.js';
