+++ assets/js/components/resource-card/resource-card.js
