+++ assets/js/components/theme-control/theme-control.js
