+++ assets/js/components/theme-slider/theme-slider.js
